moved custom review input to a seperate file
